get NFT example working
